+++ changelog.md
@@ -24,6 +24,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - shell.CommandLineToArgv
 - cmdline_from_args
 
+### Changed
+
+- virtdisk.OpenVirtualDisk returns context manager KHANDLE
+
 ## [1.3.0]
 
 ### Added